Fix TypeScript issues, and add prepare and build command
